Install dein only if ~/.vim/dein was not found

+++ vim/installer.sh
@@ -2,8 +2,10 @@
 ln -sf $PWD/vim/.vimrc ~/.vimrc
 
 # Dein.vimインストール（プラグインインストールはVim上で :call dein#install() 実行）
-curl https://raw.githubusercontent.com/Shougo/dein.vim/master/bin/installer.sh > dein.sh
-sh ./dein.sh ~/.vim/dein
-rm dein.sh
+if [ ! -e ~/.vim/dein ]; then
+  curl https://raw.githubusercontent.com/Shougo/dein.vim/master/bin/installer.sh > dein.sh
+  sh ./dein.sh ~/.vim/dein
+  rm dein.sh
+fi
 
 echo 👍 Vim setting is done!
